requirement.txt
fastapi==0.82.0
uvicorn==0.18.3
aiomysql==0.1.1
anyio==3.6.1
bcrypt==4.0.0
click==8.1.3
cryptography==38.0.1
databases==0.6.1
ecdsa==0.18.0
mysql-connector==2.2.9
mysql-connector-python==8.0.29
passlib==1.7.4
protobuf==3.20.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
SQLAlchemy==1.4.41
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
